requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Template and Static Files
jinja2==3.1.2
aiofiles==23.2.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation & ORM
pydantic==2.5.0
sqlalchemy==2.0.23

# Database Drivers
psycopg2-binary==2.9.9

# Caching & Sessions
redis==5.0.1

# WebSocket Support
websockets==12.0

# HTTP Client for Testing
requests==2.31.0

# Additional Dependencies for Production
# python-json-logger==2.0.7  # For structured logging
# prometheus-client==0.19.0  # For metrics
# gunicorn==21.2.0           # For production WSGI server